rename RuleEditorFormState to RuleEditorFormModel
